move code, cleanup
